move tests for get_surface_backend from test_surf_plotting to test_surf_utils
